fix(trace): do not call tracing.stopChunk() twice
